add topbar and footer back in, add logo
add textoverlay back in later
